+++ Funcky/Extensions/EnumerableExtensions.cs
@@ -14,7 +14,7 @@ namespace Funcky.Extensions
         /// </summary>
         /// <typeparam name="T">Type of the object.</typeparam>
         /// <returns>An <see cref="IEnumerable{T}" /> consisting of a single item or zero items.</returns>
-        public static IEnumerable<T> Yield<T>(this T? item, RequireClass<T>? ω = null)
+        public static IEnumerable<T> ToEnumerable<T>(this T? item, RequireClass<T>? ω = null)
             where T : class
         {
             if (item is { })
@@ -23,15 +23,15 @@ namespace Funcky.Extensions
             }
         }
 
-        /// <inheritdoc cref="Yield{T}(T, RequireClass{T})"/>
-        public static IEnumerable<T> Yield<T>(this T item, RequireStruct<T>? ω = null)
+        /// <inheritdoc cref="ToEnumerable{T}(T, RequireClass{T})"/>
+        public static IEnumerable<T> ToEnumerable<T>(this T item, RequireStruct<T>? ω = null)
             where T : struct
         {
             yield return item;
         }
 
-        /// <inheritdoc cref="Yield{T}(T, RequireClass{T})"/>
-        public static IEnumerable<T> Yield<T>(this T? item)
+        /// <inheritdoc cref="ToEnumerable{T}(T, RequireClass{T})"/>
+        public static IEnumerable<T> ToEnumerable<T>(this T? item)
             where T : struct
         {
             if (item.HasValue)