+++ changelog.md
@@ -13,6 +13,11 @@
 * Match of `Result` Monad accepts actions
 * Add `FirstOrNone`, `LastOrNone` and `SingleOrNone` extension functions
 
+## Unreleased
+* Added overload for AndThen which flattens the Option
+* Add `Where` method to `Option<T>`, which allows filtering the `Option` by a predicate.
+* Add overload for `Option<T>.SelectMany` that takes only a selector.
+
 ## Next major version
 * Full nullable support introduced with C# 8 
 * Rename `Option.From` -> `Option.FromNullable` and remove overload that takes non-nullable value types.
@@ -20,4 +25,5 @@
   because most APIs in `System` use `Func`.
 * `Functional.Any` now returns `false` when the given list of predicates is empty.
 * The `Each` extension method on `IEnumerable<T>` has been renamed to `ForEach`.
+* Unify `Option<T>.ToEnumerable` and `Yield` to `ToEnumerable`
 * Remove `Reader` monad based on `await`.